Included Ads

+++ app/src/main/java/com/theapphideaway/quickweather/Model/Current/Clouds.kt
@@ -1,5 +1,5 @@
 package com.theapphideaway.quickweather.Model.Current
 
 data class Clouds(
-    val all: Int
+    val all: Double
 )
