+++ app/src/main/java/com/theapphideaway/quickweather/Model/Current/Main.kt
@@ -1,8 +1,8 @@
 package com.theapphideaway.quickweather.Model.Current
 
 data class Main(
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val temp: Double,
     val temp_max: Double,
     val temp_min: Double
